refactor: update wallet connector and session management
