+++ web-v2/.env.example
@@ -3,8 +3,8 @@
 
 VITE_API_URL=http://localhost:80
 VITE_NETWORK=liquid
-VITE_ESPLORA_BASE_URL=https://blockstream.info/liquid/api
-VITE_WATERFALLS_URL=https://waterfalls.liquidwebwallet.org/liquid/api
+VITE_ESPLORA_BASE_URL=https://blockstream.info/liquid
+VITE_WATERFALLS_URL=https://waterfalls.liquidwebwallet.org/liquid
 VITE_WATERFALLS_RECIPIENT=age1xxzrgrfjm3yrwh3u6a7exgrldked0pdauvr3mx870wl6xzrwm5ps8s2h0p
 
 # Optional: BIP39 mnemonic for debug software signer (dev/testnet only).
