+++ web-v2/.env.liquid
@@ -1,5 +1,5 @@
-VITE_API_URL=http://localhost:8000
+VITE_API_URL=http://localhost:80
 VITE_NETWORK=liquid
-VITE_ESPLORA_BASE_URL=https://blockstream.info/liquid/api
-VITE_WATERFALLS_URL=https://waterfalls.liquidwebwallet.org/liquid/api
+VITE_ESPLORA_BASE_URL=https://blockstream.info/liquid
+VITE_WATERFALLS_URL=https://waterfalls.liquidwebwallet.org/liquid
 VITE_WATERFALLS_RECIPIENT=age1xxzrgrfjm3yrwh3u6a7exgrldked0pdauvr3mx870wl6xzrwm5ps8s2h0p
