+++ web-v2/.env.liquidtestnet
@@ -1,7 +1,7 @@
-VITE_API_URL=http://localhost:8000
+VITE_API_URL=http://localhost:80
 VITE_NETWORK=liquidtestnet
-VITE_ESPLORA_BASE_URL=https://blockstream.info/liquidtestnet/api
-VITE_WATERFALLS_URL=https://waterfalls.liquidwebwallet.org/liquidtestnet/api
+VITE_ESPLORA_BASE_URL=https://blockstream.info/liquidtestnet
+VITE_WATERFALLS_URL=https://waterfalls.liquidwebwallet.org/liquidtestnet
 VITE_WATERFALLS_RECIPIENT=age1xxzrgrfjm3yrwh3u6a7exgrldked0pdauvr3mx870wl6xzrwm5ps8s2h0p
 
 # Uncomment for debug software signer (dev only):
