--- web-v2/src/lib/wallet-core/wallet/session.ts
@@ -1,10 +1,0 @@
-import type { EsploraClient, Wollet, WolletDescriptor } from 'lwk_web'
-
-import type { WalletConnector } from '../connector/types'
-
-export type WalletSession = {
-  connector: WalletConnector
-  descriptor: WolletDescriptor
-  wollet: Wollet
-  esploraClient: EsploraClient
-}
